ido: add an audio hal just for a plug
so that it doesn't interfere with android loading.

for audio to work on kernel 4.9, you need to understand the driver and register it in DT. So far, this is a difficult chore that I don't want to do.

There are many other problems in the 4.9 kernel that need to be dealt with first, otherwise working with sound will not make any sense. Such problems include: cpufreq, lpm, saw - everything related to energy savings does not work, the core panics when using the Internet, and the panel also has some problems when working.