ido: display hal
these fixes are necessary so that the hal display can be assembled. This hal has been working since kernel 3.18. There are problems that I do not yet understand, they are related to the driver or to the hal itself, but one thing is clear for sure, the panel backlight does not work and when the screen is turned off, it continues to light up, otherwise, in principle, the panel works fine, sometimes it does not turn on after a long sleep.